backend/requirement.txt
flask==2.0.1
pymongo==4.6.1
python-dotenv==1.0.0
bson==0.5.10